Rename text-data.ts model file
